+++ src/Stormancer.Plugins/Users/Stormancer.Server.Plugins.Users/UsersConstants.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Stormancer.Server.Plugins.Users
+{
+    /// <summary>
+    /// Constants related to the users plugin.
+    /// </summary>
+    public static class UsersConstants
+    {
+        /// <summary>
+        /// Path to the user pseudo/handle in UserData.
+        /// </summary>
+        public const string UserHandleKey = "handle";
+    }
+}