+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -9,6 +9,16 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+4.0.7
+Changed
+*******
+- Use Stormancer.Abstraction.Server 4.2.0.
+4.0.6
+-----
+Changed
+*******
+- Fixed timeout issue with gamesession registrations to gamefinder when the gamefinder sent team updates.
+
 4.0.4.1
 -------
 Changed