+++ src/Stormancer.Plugins/Leaderboards/Stormancer.Server.Plugins.Leaderboards/Changelog.rst
@@ -9,7 +9,9 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
-
+3.0.2
+-----
+- Use Stormancer.Abstraction.Server 4.2.0.2
 3.0.1
 -----
 - Use Stormancer.Abstraction.Server 6.1.0